move & reorg all user views
